Fix current page not being updated on deleting portraits
Resets `indexPage` (current page) to `indexPageCount` (max pages) on UI
update. This will make the current page go back one if let's say the
user deleted all the portraits on page 3, there the UI will move over to
page 2 with the last portrait of that page selected.

And `manxCount` which I think is about how many portrait icons to show
for the current page was so utterly wrong it's baffling.