Fix elements getting offset to left on repeatedly opening Favorites UI
Since OpenUI for UIModelPro just brings it into focus and doesn't reset
it or something, the InitData after that will keep applying over the
previous state, and that means the offsetting of textTitle and some
other components will compound over each other.

So, for this reason, wherever there's a button that spawns this UI,
always close the UI if it's open and only then reopen it again to make
sure the state is reset.