Fix btnApply being interactable with mode 3 due to logic override
btnApply was supposed to be non-interactive under the reasoning that it
adds confusion when both "Edit" and it does the same thing. (Actually it
doesn't true, both have different semantics and uses something that'll
be fixed later)

But, with 1978235 that logic gets overridden by setting btnApply
interactibility based on selected portrait's gender.