Re-enable Portrait UI's btnApply for ModDress
It was disabled back in 736c3def2a5df38d91888d1e1c51a7fa190c2ff6 with
the concern that "Apply" only adds to the confusion with both "Edit" and
"Apply" doing the same thing for ModDress, but after reflecting on it,
they do have different semantics and usecases, so it'll be brought back.

For ModDress, "Edit" is for editing the selected portrait while "Apply"
is just to use the selected portrait in the ModDress while editing the
previously selected one, useful if you want to edit a portrait to be
similar to one of the other portraits but with slight changes, etc.

Confirmation popups for "Apply" will not appear in case of ModDress (or
UICreatePlayer).