Fix gender of portrait being stuck in ModDress after an initial btnEdit
If the user selected a male portrait for edit for the first time, and
then if they tried to select a female portrait for edit without closing
and reopening in the ModDressUI, the portrait and the outfit will switch
back to that of the initial gender on changing any dress item.

This fixes that by just closing and re-opening the UI with initial data
so the in-game logic can populate the available dress items, etc by
itself.