Prevent portrait index for edit from going invalid on deleting portraits
If user deletes the portrait that is currently set to be edited in the
ModDress or some portrait before the one to be edited, in both the cases
the index for the portrait is no longer valid, the portrait either no
longer exists or the index is pointing to some other portrait instead.

Fix this by keeping track of the index that is to be edited and adjust
it accordingly based on which portrait is to be deleted while at the
same time disallow deletion of the portrait that is to be edited.