Fix Portraits UI not getting updated after "Edit Complete" in some cases
When the instance of Portraits UI that spawned the UIModDress goes away,
because maybe it got reopened, that instance is no longer valid and the
callback to "Edit Complete" calls UpData() on that very instance which
obviously doesn't work.

Fix this by getting the latest UIBase if present using `g.ui.GetUI` and
performing UpData() on its UIModelPro component.