Fix nonexistent key errors for dataSpiritTalents due to locked Spirit
This happens because the Artifact Spirit may not be unlocked at the
time, and so the `Spirit` isn't there along with its corresponding
`List<Talent>`, hence the missing keys.

Now, there's no need to rebuild `dataSpiritTalents` and apply the
modifications again on player unlocking a new Artifact Spirit after the
IntoWorld event because I think the Talent's `unlock(2|3)Count` field's
value should be taken from `ConfArtifactSpriteTalent`, something that
we've already modified.