Fix "No Cost" failure popup on talent unlock for newly acquired Spirits
Can be fixed by rebuilding dataSpiritTalents on opening of UIArtificat,
or by rebuilding it only on acquiring a new Spirit by hooking into the
UI that pops up when you unlock a new Spirit in Fallen Valley and check
if it's a new one. The former approach is taken in this commit.